feat(components global): add doc
