+++ src/ui-example/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../commons/styles/style.scss">
+    <title>Document</title>
+</head>
+<body>
+    <div class="test">
+
+        <h2>H2</h2>
+        <p>h2 ou la classe h2-title</p>
+        <h3>H3</h3>
+        <p>h3 ou la classe h3-title</p>
+        <h3 class="h3-title-serif">H3</h3>
+        <p>la classe h3-title-serif</p>
+        <p>Paragraph</p>
+        <p>p ou la classe content </p>
+        
+
+    </div>
+    <script src="./useComponents.js" type="module"></script>
+</body>
+</html>
